Saved every tenth checkpoint with optimizer

+++ code/src/pipeline/evaluator.py
@@ -407,14 +407,15 @@ def main(config_filename: str):
                             print(f"Delete: {os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint)}")
                             shutil.rmtree(os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint))
                             # Delete model with optimizer:
-                            opt_dir = os.path.join(MODEL_CHECKPOINT_PATH, "optimizer")
-                            selected_files = []
-                            for f in os.listdir(opt_dir):
-                                if f.startswith(unevaluated_checkpoint):
-                                    selected_files.append(f)
+                            if int(unevaluated_checkpoint[5:]) % 10 == 0:
+                                opt_dir = os.path.join(MODEL_CHECKPOINT_PATH, "optimizer")
+                                selected_files = []
+                                for f in os.listdir(opt_dir):
+                                    if f.startswith(unevaluated_checkpoint):
+                                        selected_files.append(f)
 
-                            for selected_file in selected_files:
-                                os.remove(os.path.join(opt_dir, selected_file))
+                                for selected_file in selected_files:
+                                    os.remove(os.path.join(opt_dir, selected_file))
                 except Exception as e:
                     print(e)
                     print("Something went wrong... Try again...")
